clip-server/requirements.txt
fastapi
uvicorn
torch
clip-by-openai
pymongo
pillow
nltk
scikit-learn
python-dotenv
numpy 